Explicit boxing
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1724693 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 6af9d4bd6ba83b75a7f6920eea0b952e08f3c598

+++ src/jorphan/org/apache/jorphan/util/JOrphanUtils.java
@@ -573,6 +573,6 @@ public final class JOrphanUtils {
      * @return String formated with format HH:mm:ss
      */
     public static String formatDuration(long elapsedSec) {
-        return String.format("%02d:%02d:%02d", elapsedSec / 3600, (elapsedSec % 3600) / 60, (elapsedSec % 60));
+        return String.format("%02d:%02d:%02d", Long.valueOf(elapsedSec / 3600), Long.valueOf((elapsedSec % 3600) / 60), Long.valueOf(elapsedSec % 60));
     }
 }
